+++ src/components/GInput/GInputField.js
@@ -119,11 +119,6 @@ export function getEvents(props, context, internalValue, isFocused, isValidInput
     }
   }
 
-  function onInput() {
-    context.emit('input', event);
-    props.autoGrow && calculateInputHeight(props,context)
-  }
-
   function onBlur(event) {
     context.emit('blur', event);
     if(props.validateOnBlur)
@@ -134,13 +129,13 @@ export function getEvents(props, context, internalValue, isFocused, isValidInput
   }
 
   function onClearIconClick(event) {
-    console.log('ok')
     internalValue.value = ''
     isValidInput.value = true;
     context.emit('click:clearIcon', event)
   }
 
   function onChange(event) {
+    console.log('change')
     context.emit('change', event)
   }
 
@@ -172,7 +167,7 @@ export function getEvents(props, context, internalValue, isFocused, isValidInput
     }
   }
 
-  return { onInput, onClick, onFocus, onBlur, onClearIconClick, onMouseDown, onMouseUp, onChange, onKeyDown }
+  return { onClick, onFocus, onBlur, onClearIconClick, onMouseDown, onMouseUp, onChange, onKeyDown }
 }
 
 export function getInternalValue(props, context) {
@@ -190,16 +185,4 @@ export function getInternalValue(props, context) {
   });
 
   return internalValue;
-}
-
-export function calculateInputHeight(props, context) {
-  const input = context.refs.input
-  if (!input) return
-
-  input.style.height = '0'
-  const height = input.scrollHeight
-  const minHeight = parseInt(props.rows, 10) * parseFloat(props.rowHeight)
-  // This has to be done ASAP, waiting for Vue
-  // to update the DOM causes ugly layout jumping
-  input.style.height = Math.max(minHeight,height) + 10 + 'px'
 }
