--- src/components/GTextarea/GTear.vue
@@ -1,275 +1,0 @@
-<template>
-  <div class="tf-wrapper" :class="[tfWrapperClasses, tfErrWrapperClass]" @click="onClick" @mouseup="onMouseUp"
-       @mousedown="onMouseDown">
-    <div class="tf-prepend__outer" ref="prependRef" @click="onClickPrependOuter">
-      <slot name="prepend-outer"></slot>
-    </div>
-    <fieldset>
-      <legend :style="legendStyles">{{label}}</legend>
-      <div class='tf' :class="tfErrClasses">
-        <div class="tf-prepend__inner" @click="onClickPrependInner">
-          <slot name="prepend-inner"></slot>
-        </div>
-        <div v-if="prefix" class="tf-affix" ref="prefixRef">{{prefix}}</div>
-        <div class="inputGroup">
-          <!--          <input id="input" type="text"-->
-          <!--                 class="tf-input"-->
-          <!--                 :style="inputErrStyles"-->
-          <!--                 :type="type"-->
-          <!--                 :label="label"-->
-          <!--                 v-model="internalValue"-->
-          <!--                 :placeholder="placeholder"-->
-          <!--                 ref="input"-->
-          <!--                 @change="onChange"-->
-          <!--                 @focus="onFocus"-->
-          <!--                 @blur="onBlur"-->
-          <!--                 @keydown="onKeyDown">-->
-          <textarea id="tear" ref="input"
-                    class="tf-input"
-                    :style="tearStyles"
-                    :label="label"
-                    :type="type"
-                    v-model="internalValue"
-                    :placeholder="placeholder"
-                    :rows="rows"
-                    :cols="cols"
-                    @input="onInput"
-                    @change="onChange"
-                    @focus="onFocus"
-                    @blur="onBlur"
-                    @keydown="onKeyDown">
-          </textarea>
-          <label for="tear" class="tf-label" :style="labelStyles">
-            <slot name="label">{{label}}</slot>
-          </label>
-        </div>
-        <div v-if="suffix" class="tf-affix">{{suffix}}</div>
-        <div class="tf-append__inner" @click="onClickAppendInner">
-          <slot name="append-inner"></slot>
-        </div>
-        <div class="tf-error" v-if="!isValidInput">{{errorMessages}}</div>
-        <div class="tf-hint" v-else :class="hintClasses">{{hint}}</div>
-        <div v-show="counter" :class="{'tf-counter': true, 'tf-counter__error': !isValidInput}">{{internalValue.length}}
-          / {{counter}}
-        </div>
-      </div>
-    </fieldset>
-    <div class="tf-append__outer" @click="onClickAppendOuter" ref="appendOuter">
-      <!--      <img v-if="isDirty && clearable" src="../../assets/delivery/cancel.svg" @click="onClearIconClick">-->
-      <div v-if="isDirty && clearable" @click="onClearIconClick">
-        <g-icon class="g-icon__link">cancel</g-icon>
-      </div>
-      <!--      <g-icon v-if="isDirty && clearable" @click="onClearIconClick">cancel</g-icon>-->
-      <slot name="append-outer"></slot>
-    </div>
-  </div>
-</template>
-
-<script>
-  import {ref, computed, onMounted, onUpdated} from '@vue/composition-api';
-  import {
-    getEvents,
-    getInternalValue,
-    getLabel,
-    getSlotEventListeners,
-    getValidate,
-    calculateInputHeight
-  } from '../GInput/GInputField';
-  import GIcon from "../GIcon/GIcon";
-
-  export default {
-    name: 'GTextFieldFunctional',
-    components: {GIcon},
-    props: {
-      ...{//display props
-        label: String,
-        placeholder: String,
-        appendIcon: String,
-        prependIcon: String,
-        prefix: {
-          type: String,
-          default: ''
-        },
-        suffix: {
-          type: String,
-          default: ''
-        },
-        //input states
-        clearable: Boolean,
-        disabled: Boolean,
-        readOnly: Boolean,
-      },
-      //rules and validation props
-      rules: Array,
-      hint: String,
-      errorCount: {
-        type: Number,
-        default: 1
-      },
-      persistent: Boolean,
-      counter: [Number, Boolean, String],
-      validateOnBlur: Boolean,
-      error: Boolean,
-
-      //styles
-      filled: Boolean,
-      outlined: Boolean,
-      solo: Boolean,
-      shaped: Boolean,
-      rounded: Boolean,
-      flat: Boolean,
-
-      // basic props
-      value: [String, Number],
-      type: {
-        type: String,
-        default: 'text',
-      },
-
-      //todo textarea
-      autoGrow: Boolean,
-      noResize: Boolean,
-      rowHeight: {
-        type: [Number, String],
-        default: 24,
-        validator: (v) => !isNaN(parseFloat(v)),
-      },
-      rows: {
-        type: [Number, String],
-        default: 5,
-        validator: (v) => !isNaN(parseInt(v, 10)),
-      },
-      cols: {
-        type: [Number, String],
-        default: 1,
-        validator: (v) => !isNaN(parseInt(v, 10)),
-      },
-    },
-    setup(props, context) {
-      const tfWrapperClasses = getTfWrapperClasses(props);
-
-      const internalValue = getInternalValue(props, context);
-      const isValidInput = ref(true)
-      const isFocused = ref(false);
-
-      const {labelClasses, labelStyles, isDirty, isLabelActive, prefixRef} = getLabel(props, internalValue, isValidInput, isFocused, 'tf-label__active', {'color': 'red'});
-
-      //Activate non persistent hint
-      const hintClasses = computed(() => (props.persistent || (isFocused.value && isValidInput.value)) ? {'tf-hint__active': true} : {})
-
-      //event handler function
-      const {errorMessages, validate} = getValidate(props, isFocused, internalValue, isValidInput);
-
-      const inputErrStyles = computed(() => isValidInput.value ? {} : {'color': 'red'})
-      const tearStyles = computed(() => {
-        return {
-          ...inputErrStyles,
-          'line-height': props.rowHeight + 'px',
-        }
-      })
-      //change input border color
-      const tfErrClasses = computed(() => isValidInput.value ? {} : {'tf__error': true})
-
-      const tfErrWrapperClass = computed(() => ({'tf-wrapper__error': !isValidInput.value}));
-
-      const {onClickPrependOuter, onClickPrependInner, onClickAppendOuter, onClickAppendInner,} = getSlotEventListeners(context);
-
-      const {
-        onInput, onClick, onFocus, onBlur, onClearIconClick,
-        onMouseDown, onMouseUp, onChange, onKeyDown
-      } = getEvents(props, context, internalValue, isFocused, isValidInput, validate);
-      //set legend width for label in outlined textfield
-      const legendStyles = computed(() => {
-        if (isFocused.value || internalValue.value) {
-          return {
-            'width': 'auto',
-            'padding': '1px',
-          }
-        } else
-          return {}
-      });
-
-      //todo add textarea logic
-
-      onMounted(() => {
-        setTimeout(() => {
-          //props.autoGrow && calculateInputHeight(props,context)
-          calculateInputHeight(props, context)
-        }, 0)
-      })
-
-      onUpdated(() => {
-        setTimeout(() => {
-          props.autoGrow && calculateInputHeight(props, context)
-          //calculateInputHeight(props, context)
-        }, 0)
-      })
-
-      return {
-        //calculated styles and classes
-        labelClasses,
-        labelStyles,
-        tfErrClasses,
-        tfWrapperClasses,
-        hintClasses,
-        tearStyles,
-
-        //value
-        internalValue,
-
-        //calculated state
-        isLabelActive,
-        isFocused,
-        isDirty,
-        isValidInput,
-
-        //calculated error
-        errorMessages,
-
-        //event listeners
-        onInput,
-        onClick,
-        onChange,
-        onFocus,
-        onBlur,
-        onKeyDown,
-        onMouseUp,
-        onMouseDown,
-        onClickPrependOuter,
-        onClickPrependInner,
-        onClickAppendOuter,
-        onClickAppendInner,
-        onClearIconClick,
-
-        //ref
-        prefixRef,
-        tfErrWrapperClass,
-        legendStyles,
-      }
-    }
-  }
-
-  function getTfWrapperClasses(props) {
-    return computed(() => (props.disabled ? {'tf-wrapper-disabled': true} : {
-      'tf__filled': props.filled,
-      'tf__outlined': props.outlined,
-      'tf__solo': props.solo,
-      'tf__rounded': props.rounded,
-      'tf__shaped': props.shaped,
-      'tf__flat': props.flat,
-      'tf-wrapper-readonly': props.readOnly,
-      //textarea
-      'g-tear': true,
-      'g-tear__auto-grow': props.autoGrow,
-      'g-tear__no-resize': props.noResize || props.autoGrow,
-    }))
-  }
-
-</script>
-
-<style lang="scss">
-  @import "GTear";
-</style>
-
-
-
