--- src/components/GTextarea/_GTear.scss
@@ -1,471 +1,0 @@
-@import "../../style/elevation";
-
-.g-tear {
-  textarea {
-    flex: 1 1 auto;
-    line-height: 22px;
-    max-width: 100%;
-    min-height: 32px;
-    outline: none;
-    padding: 7px 0 8px;
-    width: 100%;
-    // inherit from input
-    //border: transparent;
-    //font-family: "Muli", sans-serif;
-  }
-
-  &__auto-grow {
-    textarea {
-      overflow: hidden;
-    }
-  }
-
-  &__no-resize {
-    textarea {
-      resize: none;
-    }
-  }
-}
-
-textarea[type=password] {
-  outline: none;
-  border: transparent;
-  font-family: "Muli", sans-serif;
-}
-
-textarea[type=text] {
-  outline: none;
-  border: transparent;
-  font-family: "Muli", sans-serif;
-}
-
-.inputGroup {
-  position: relative;
-  width: 100%;
-  left: 5px;
-}
-
-.tf-wrapper {
-  display: flex;
-  align-items: center;
-  cursor: text;
-  width: calc(100% - 10px);
-  margin: 16px 5px 24px;
-  position: relative;
-
-  &__error {
-    color: red;
-  }
-
-  fieldset {
-    width: 100%;
-    border: 0;
-    padding: 0;
-    margin: 0;
-  }
-
-  legend {
-    display: none;
-  }
-
-  &-disabled {
-    pointer-events: none;
-
-    .tf-input,
-    .tf-label {
-      opacity: 0.5;
-    }
-  }
-
-  &-readonly {
-    pointer-events: none;
-  }
-}
-
-.tf {
-  position: relative;
-  flex: 1 1 auto;
-  color: rgba(0, 0, 0, 0.32);
-  display: flex;
-
-  &::before {
-    position: absolute;
-    content: '';
-    bottom: -1px;
-    width: 100%;
-    height: 1px;
-    background-color: rgba(0, 0, 0, 0.87);
-  }
-
-  &::after {
-    position: absolute;
-    content: '';
-    bottom: -1px;
-    width: 100%;
-    height: 1px;
-    background-color: #1867c0;
-    transform: scale(0);
-    transition: 0.4s ease-out;
-  }
-
-  &:focus-within::after {
-    height: 2px;
-    transform: scale(1);
-  }
-
-  &__error::before {
-    background-color: rgba(192, 2, 19, 0.87);
-  }
-
-  &__error::after {
-    background-color: #c00213;
-  }
-
-  &-input {
-    width: 100%;
-    padding: 12px 0;
-    line-height: 24px;
-    font-size: 16px;
-    outline: none;
-  }
-
-  &-affix {
-    padding: 12px 0 4px 12px;
-    line-height: 24px;
-    font-size: 16px;
-  }
-
-  &-label {
-    position: absolute;
-    top: 12px;
-    left: 0;
-    transform-origin: top left;
-    transition: 0.4s ease-out;
-    pointer-events: none;
-    color: #1867c0;
-    font-size: 16px;
-
-    &__active {
-      transform: translateY(-18px) scale(0.75);
-      background-color: white;
-    }
-  }
-
-  &-hint {
-    position: absolute;
-    bottom: -22px;
-    opacity: 0;
-    left: 0;
-    font-size: 12px;
-    font-weight: 400;
-    transition: 0.4s ease-out;
-    line-height: 1.34;
-
-    &__active {
-      opacity: 1;
-    }
-  }
-
-  &-error,
-  &-counter {
-    position: absolute;
-    bottom: -18px;
-    font-size: 12px;
-    font-weight: 400;
-    line-height: 1.34;
-  }
-
-  &-error {
-    left: 0;
-    color: red;
-  }
-
-  &-counter {
-    right: 0;
-
-    &__error {
-      color: red;
-    }
-  }
-}
-
-.tf-prepend__outer,
-.tf-append__outer {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 0 4px;
-}
-
-.tf-prepend__inner,
-.tf-append__inner {
-  display: flex;
-  align-items: flex-end;
-  justify-content: center;
-  margin-bottom: 4px;
-}
-
-.tf-icon__active {
-  flex: 0 0 4%;
-}
-
-.tf-prepend__inner.tf-icon__active ~ .tf-label {
-  left: 4%;
-}
-
-
-.tf__solo {
-  &.tf-wrapper .tf {
-    border-radius: 4px;
-    @include elevation(2);
-  }
-
-  .tf::before,
-  .tf::after,
-  .tf-label {
-    display: none;
-  }
-
-  .tf-input {
-    padding: 8px;
-    border-radius: 4px;
-  }
-
-  .tf-hint,
-  .tf-error {
-    left: 4px;
-  }
-}
-
-.tf__filled {
-  .tf {
-    background-color: rgba(0, 0, 0, 0.06);
-    border-top-left-radius: 4px;
-    border-top-right-radius: 4px;
-  }
-
-  .tf-input {
-    background-color: transparent;
-    border: inherit;
-    padding: 12px 12px 4px;
-    line-height: 32px;
-    vertical-align: bottom;
-  }
-
-  .tf-label {
-    left: 7px;
-    top: 16px;
-  }
-
-  .tf-input:focus + .tf-label,
-  .tf-input:placeholder-shown + .tf-label {
-    transform: translateY(-16px) scale(0.75);
-    color: #1867c0;
-  }
-
-  .tf-label__active {
-    transform: translateY(-16px) scale(0.75);
-    background-color: transparent;
-  }
-
-  .tf-affix {
-    line-height: 32px;
-  }
-
-  .tf-prepend__inner.tf-icon__active ~ .tf-label {
-    left: calc(4% + 4px);
-  }
-
-  .tf-hint,
-  .tf-error {
-    left: 12px;
-  }
-}
-
-.tf__outlined {
-  &.tf-wrapper fieldset {
-    display: flex;
-    flex: 1 1 auto;
-    position: relative;
-    border: 1px solid rgba(0, 0, 0, 0.87);
-    border-radius: 4px;
-    padding: 4px 8px;
-    transition: 0.2s ease-in;
-    width: 100%;
-  }
-
-  &.tf-wrapper__error fieldset {
-    border-color: red;
-  }
-
-  &.tf-wrapper legend {
-    display: block;
-    width: 0;
-    height: 0;
-    color: transparent;
-    font-size: 0.75em;
-    padding: 0;
-    transition: 0.8s cubic-bezier(.44, .88, .37, .28);
-  }
-
-  .tf::before,
-  .tf::after {
-    display: none;
-  }
-
-  &:focus-within fieldset {
-    border: 2px solid #1867c0;
-  }
-
-  &.tf-wrapper__error:focus-within fieldset {
-    border-color: red;
-  }
-
-  &:focus-within legend {
-    width: auto;
-    padding: 1px;
-  }
-
-  .tf-prepend__inner.tf-icon__active ~ .tf-label {
-    left: calc(4% + 8px);
-  }
-
-  .tf-prepend__inner.tf-icon__active ~ .tf-label__active {
-    left: 0;
-  }
-
-  &:focus-within .tf-prepend__inner.tf-icon__active ~ .tf-label {
-    left: 0;
-  }
-
-  .tf-input {
-    border-radius: inherit;
-  }
-
-  .tf-hint,
-  .tf-error {
-    left: 12px;
-  }
-}
-
-.tf__rounded {
-  &.tf__solo .tf,
-  &.tf__outlined .tf,
-  &.tf__filled .tf {
-    border-radius: 9999px;
-  }
-
-  &.tf__solo .tf-input {
-    margin: 0 16px;
-  }
-
-  &.tf__filled .tf::before,
-  &.tf__filled .tf::after {
-    display: none;
-  }
-
-  &.tf__outlined fieldset {
-    border-radius: 9999px;
-  }
-
-  &.tf__outlined legend {
-    margin-left: 8px;
-  }
-
-  &.tf__filled .tf-label {
-    left: 16px;
-  }
-
-  &.tf__outlined .tf-label {
-    top: 4px;
-    left: 8px;
-    transform: translateY(4px);
-  }
-
-  &.tf__outlined:focus-within .tf-label {
-    left: 8px;
-  }
-
-  &.tf__filled .tf-input {
-    padding-left: 16px;
-  }
-
-  &.tf__outlined .tf-input {
-    padding-left: 8px;
-  }
-
-  &.tf__solo .tf-prepend__outer.tf-icon__active ~ .tf-hint,
-  &.tf__outlined .tf-prepend__outer.tf-icon__active ~ .tf-hint,
-  &.tf__filled .tf-prepend__outer.tf-icon__active ~ .tf-hint {
-    left: calc(4% + 16px);
-  }
-}
-
-.tf__shaped {
-  &.tf__solo .tf,
-  &.tf__outlined .tf,
-  &.tf__filled .tf {
-    border-top-left-radius: 16px;
-    border-top-right-radius: 16px;
-  }
-
-  &.tf__solo .tf-input {
-    margin: 0 16px;
-  }
-
-  &.tf__filled .tf::before,
-  &.tf__filled .tf::after {
-    display: none;
-  }
-
-  &.tf__outlined fieldset {
-    border-top-left-radius: 16px;
-    border-top-right-radius: 16px;
-  }
-
-  &.tf__filled .tf-label {
-    left: 16px;
-  }
-
-  &.tf__outlined .tf-label {
-    top: 4px;
-    transform: translateY(4px);
-    left: 8px;
-  }
-
-  &.tf__outlined:focus-within .tf-label {
-    left: 8px;
-  }
-
-  &.tf__filled .tf-input {
-    padding-left: 16px;
-  }
-
-  &.tf__outlined .tf-input {
-    padding-left: 8px;
-  }
-
-  &.tf__solo .tf-prepend__outer.tf-icon__active ~ .tf-hint,
-  &.tf__outlined .tf-prepend__outer.tf-icon__active ~ .tf-hint,
-  &.tf__filled .tf-prepend__outer.tf-icon__active ~ .tf-hint {
-    left: calc(4% + 16px);
-  }
-}
-
-.tf-wrapper.tf__solo.tf__outlined legend {
-  width: 0;
-}
-
-.tf-wrapper.tf__solo.tf__outlined .tf {
-  box-shadow: none;
-}
-
-.tf-wrapper.tf__solo.tf__outlined {
-  border-radius: 4px;
-  box-shadow: 0px 3px 1px -2px rgba(0, 0, 0, 0.2), 0px 2px 2px 0px rgba(0, 0, 0, 0.14), 0px 1px 5px 0px rgba(0, 0, 0, 0.12);
-}
-
-.tf-wrapper.tf__solo.tf__flat .tf {
-  border: none;
-  box-shadow: none;
-}
