+++ src/style/main.scss
@@ -28,5 +28,5 @@
 @import "../components/GDialog/_GDialog";
 @import "../components/GImg/_GImg";
 @import "../components/GIcon/_GIcon";
-@import "../components/GTextarea/GTear";
+@import "../components/GTextarea/GTextarea";
